requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
oauth2client==4.1.3
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyparsing==3.1.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
